+++ src/main/java/be/ugent/idlab/knows/dataio/record/JSONRecord.java
@@ -67,7 +67,7 @@ public class JSONRecord extends Record {
         }
 
         if (!reference.contains("$")) {
-            reference = String.format("$.%s", reference);
+            reference = reference.startsWith(".") ? String.format("$%s", reference) : String.format("$.%s", reference);
         }
 
         if (reference.equals("@")) {